bench: expand benchmarks to 31 total (+14 new)
New benchmarks:
- MC barrier (down-and-out 50k paths)
- MC Asian (arithmetic average 50k paths)
- MC Heston (50k paths, 252 steps)
- MC Bates / Heston+jumps (50k paths)
- FD Heston 2D Douglas ADI (50×30×50)
- LSM American put (50k paths, Laguerre basis)
- Kirk spread call approximation
- MC basket (3-asset correlated, 50k paths)
- SABR implied volatility
- SVI calibration (9 strikes)
- HW Jamshidian swaption (10y swap)
- Merton jump-diffusion call
- Bates analytic (flat params)
- Schedule generation (30y semiannual)

Also fixed:
- American approx benchmarks: OptionType::Put → bool false
- Pre-existing clippy: unnecessary i32 cast in cap_floor_engines

910 tests pass, 0 clippy warnings (lib)

+++ crates/ql-pricingengines/src/cap_floor_engines.rs
@@ -148,7 +148,7 @@ mod tests {
                 accrual_fraction: 0.5,
                 notional: 1_000_000.0,
                 forward_rate,
-                discount: 0.97_f64.powi(i as i32 + 1),
+                discount: 0.97_f64.powi(i + 1),
             })
             .collect()
     }
